importing App class into main.rb
